cleaned up demos
