+++ examples/__init__.py
@@ -1,4 +1,4 @@
-from .new_project_templates.lightning_module_template import LightningTemplateModel
+from .basic_examples.lightning_module_template import LightningTemplateModel
 
 __all__ = [
     'LightningTemplateModel'
