+++ examples/basic_examples/lightning_module_template.py
