+++ examples/domain_templates/gan.py
