+++ examples/multi_node_examples/demo_script.sh
