+++ examples/multi_node_examples/minimal_multi_node_demo_script.sh
